third_party/blink/web_tests/external/wpt/web-animations/animation-model/animation-types/interpolation-per-property-002-expected.txt

This is a testharness.js-based test.
Found 13 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] list-style uses discrete animation when animating between "inside url("http://localhost/test-1") circle" and "outside url("http://localhost/test-2") square" with linear easing
  assert_equals: The value should be inside url("http://localhost/test-1") circle at 499ms expected "inside url(\\"http://localhost/test-1\\") circle" but got "inside cross-fade(url(\\"http://localhost/test-1\\") 50.1%, url(\\"http://localhost/test-2\\")) circle"
[FAIL] list-style uses discrete animation when animating between "inside url("http://localhost/test-1") circle" and "outside url("http://localhost/test-2") square" with effect easing
  assert_equals: The value should be inside url("http://localhost/test-1") circle at 940ms expected "inside url(\\"http://localhost/test-1\\") circle" but got "inside cross-fade(url(\\"http://localhost/test-1\\") 57.1741%, url(\\"http://localhost/test-2\\")) circle"
[FAIL] list-style uses discrete animation when animating between "inside url("http://localhost/test-1") circle" and "outside url("http://localhost/test-2") square" with keyframe easing
  assert_equals: The value should be inside url("http://localhost/test-1") circle at 940ms expected "inside url(\\"http://localhost/test-1\\") circle" but got "inside cross-fade(url(\\"http://localhost/test-1\\") 57.1741%, url(\\"http://localhost/test-2\\")) circle"
[FAIL] list-style-image uses discrete animation when animating between "url("http://localhost/test-1")" and "url("http://localhost/test-2")" with linear easing
  assert_equals: The value should be url("http://localhost/test-1") at 499ms expected "url(\\"http://localhost/test-1\\")" but got "cross-fade(url(\\"http://localhost/test-1\\") 50.1%, url(\\"http://localhost/test-2\\"))"
[FAIL] list-style-image uses discrete animation when animating between "url("http://localhost/test-1")" and "url("http://localhost/test-2")" with effect easing
  assert_equals: The value should be url("http://localhost/test-1") at 940ms expected "url(\\"http://localhost/test-1\\")" but got "cross-fade(url(\\"http://localhost/test-1\\") 57.1741%, url(\\"http://localhost/test-2\\"))"
[FAIL] list-style-image uses discrete animation when animating between "url("http://localhost/test-1")" and "url("http://localhost/test-2")" with keyframe easing
  assert_equals: The value should be url("http://localhost/test-1") at 940ms expected "url(\\"http://localhost/test-1\\")" but got "cross-fade(url(\\"http://localhost/test-1\\") 57.1741%, url(\\"http://localhost/test-2\\"))"
[FAIL] math-depth supports animating as an integer
  assert_equals: The value should be 0 at 500ms expected "0" but got "2"
[FAIL] position-visibility uses discrete animation when animating between "always" and "anchors-valid" with linear easing
  assert_equals: The value should be anchors-valid at 500ms expected "anchors-valid" but got "anchors-visible"
[FAIL] position-visibility uses discrete animation when animating between "always" and "anchors-valid" with effect easing
  assert_equals: The value should be anchors-valid at 960ms expected "anchors-valid" but got "anchors-visible"
[FAIL] position-visibility uses discrete animation when animating between "always" and "anchors-valid" with keyframe easing
  assert_equals: The value should be anchors-valid at 960ms expected "anchors-valid" but got "anchors-visible"
[FAIL] text-decoration uses discrete animation when animating between "underline" and "overline" with linear easing
  assert_equals: The value should be underline at 0ms expected "underline" but got "underline rgb(0, 0, 0)"
[FAIL] text-decoration uses discrete animation when animating between "underline" and "overline" with effect easing
  assert_equals: The value should be underline at 0ms expected "underline" but got "underline rgb(0, 0, 0)"
[FAIL] text-decoration uses discrete animation when animating between "underline" and "overline" with keyframe easing
  assert_equals: The value should be underline at 0ms expected "underline" but got "underline rgb(0, 0, 0)"
Harness: the test ran to completion.
